Move heavy top related logic to separate source files
