+++ src/rigid_pendulum_poc/CMakeLists.txt
@@ -4,6 +4,7 @@ target_sources(${oturb_lib_name}
     # diagnostics.cpp
     # io.cpp
     generalized_alpha_time_integrator.cpp
+    heavy_top.cpp
     quaternion.cpp
     solver.cpp
     state.cpp
