+++ src/rigid_pendulum_poc/generalized_alpha_time_integrator.h
@@ -92,14 +92,4 @@ private:
     HostView2D ComputeIterationMatrix(HostView1D, std::function<HostView2D(size_t)>);
 };
 
-HostView2D heavy_top_iteration_matrix(
-    HostView2D, HostView2D, HostView2D, HostView1D, HostView1D, HostView1D, const double,
-    const double
-);
-HostView2D heavy_top_tangent_damping_matrix(HostView1D, HostView2D);
-HostView2D heavy_top_tangent_stiffness_matrix(HostView1D, HostView2D, HostView1D);
-HostView2D heavy_top_constraint_gradient_matrix(HostView1D, HostView2D);
-
-HostView2D rigid_pendulum_iteration_matrix(size_t size);
-
 }  // namespace openturbine::rigid_pendulum
